+++ libsrc/leddevice/dev_net/ProviderUdp.h
@@ -28,6 +28,8 @@ public:
 	///
 	~ProviderUdp() override;
 
+	QHostAddress getAddress() const { return _address; }
+
 protected:
 
 	///
@@ -53,8 +55,7 @@ protected:
 	int close() override;
 
 	///
-	/// @brief Writes the given bytes/bits to the UDP-device and sleeps the latch time to ensure that the
-	/// values are latched.
+	/// @brief Writes the given bytes to the UDP-device
 	///
 	/// @param[in] size The length of the data
 	/// @param[in] data The data
@@ -62,6 +63,15 @@ protected:
 	/// @return Zero on success, else negative
 	///
 	int writeBytes(const unsigned size, const uint8_t *data);
+
+	///
+	/// @brief Writes the given bytes to the UDP-device
+	///
+	/// @param[in] data The data
+	///
+	/// @return Zero on success, else negative
+	///
+	int writeBytes(const QByteArray &bytes);
 
 	///
 	QUdpSocket * _udpSocket;